requirements.txt
Flask>=2.0.0
Django>=3.0.0
FastAPI>=0.68.0
scikit-learn>=1.0.0
numpy>=1.21.0
PyYAML>=6.0
pydantic>=1.8.0
uvicorn>=0.15.0